fixing and testing to like unfinished

+++ backend/src/Application/Service/Post/PostDeleteService.php
@@ -9,8 +9,10 @@ use src\Application\Service\ServiceHelper;
 
 use src\Domain\Repository\PostRepositoryInterface;
 use src\Application\DTO\Post\PostDeleteDTO;
-
+use src\Domain\Entity\PostType;
+use src\Shared\Exception\BusinessException\BusinessException;
 use src\Shared\Exception\BusinessException\EntityNotFoundException;
+use src\Shared\Exception\BusinessException\InvalidValueException;
 
 class PostDeleteService
 {
@@ -20,15 +22,19 @@ class PostDeleteService
     {
         $post = $this->postRepo->getPostById($DTO->postToDeleteId);
 
+        if($post === null)
+            throw new EntityNotFoundException("Post", $DTO->postToDeleteId);
+
+        if(($postType = PostType::tryFrom($DTO->postType)) === null)
+            throw new InvalidValueException("postType", $DTO->postType);
+
+        ServiceHelper::validatePostType($postType, $post);
+
         ServiceHelper::authorizeAction(
             UserRole::from($DTO->loggedUserRole),
             $DTO->loggedUserId,
-            $post->getId()
+            $post->userId
         );
-
-          
-        if($post === null)
-            throw new EntityNotFoundException("Post", $DTO->postToDeleteId);
 
         $this->postRepo->deletePost($post->getId());
     }
